feat: Add profile dropdown and farmer/supplier dashboard
- I created a new profile dropdown menu that appears on hover.
- I added a new dashboard layout for farmers and suppliers with a sidebar for navigation.
- I moved the product management functionality to a new `DashboardProductList` component.
- I protected the dashboard route to only allow access for logged-in farmers and suppliers.
