+++ frontend/src/styles/Header.css
@@ -61,6 +61,7 @@
     align-items: center;
     gap: 8px;
     cursor: pointer;
+    position: relative;
 }
 
 .action-item i {
